+++ lib/Db/Comment.php
@@ -27,6 +27,7 @@ namespace OCA\Polls\Db;
 
 use JsonSerializable;
 
+use OCP\IUser;
 use OCP\AppFramework\Db\Entity;
 
 /**
@@ -76,7 +77,17 @@ class Comment extends Entity implements JsonSerializable {
 			'userId' => $this->userId,
 			'dt' => $this->dt,
 			'timestamp' => intval($timestamp),
-			'comment' => $this->comment
+			'comment' => $this->comment,
+			'displayName' => $this->getDisplayName()
 		];
+	}
+
+	private function getDisplayName() {
+
+		if (\OC::$server->getUserManager()->get($this->userId) instanceof IUser) {
+			return \OC::$server->getUserManager()->get($this->userId)->getDisplayName();
+		} else {
+			return $this->userId;
+		}
 	}
 }