+++ lib/Model/Acl.php
@@ -30,9 +30,11 @@ use OCP\AppFramework\Db\DoesNotExistException;
 
 use OCP\IGroupManager;
 use OCP\ILogger;
+use OCP\IUser;
 use OCA\Polls\Db\Poll;
 use OCA\Polls\Db\Share;
 use OCA\Polls\Db\PollMapper;
+use OCA\Polls\Db\VoteMapper;
 use OCA\Polls\Db\ShareMapper;
 
 /**
@@ -44,6 +46,7 @@ class Acl implements JsonSerializable {
 
 	/** @var int */
 	private $pollId = 0;
+
 	/** @var ILogger */
 	private $logger;
 
@@ -79,6 +82,7 @@ class Acl implements JsonSerializable {
 	 * @param ILogger $logger
 	 * @param IGroupManager $groupManager
 	 * @param PollMapper $pollMapper
+	 * @param VoteMapper $voteMapper
 	 * @param ShareMapper $shareMapper
 	 * @param Poll $pollMapper
 	 *
@@ -88,6 +92,7 @@ class Acl implements JsonSerializable {
 		ILogger $logger,
 		IGroupManager $groupManager,
 		PollMapper $pollMapper,
+		VoteMapper $voteMapper,
 		ShareMapper $shareMapper,
 		Poll $poll
 	) {
@@ -95,6 +100,7 @@ class Acl implements JsonSerializable {
 		$this->logger = $logger;
 		$this->groupManager = $groupManager;
 		$this->pollMapper = $pollMapper;
+		$this->voteMapper = $voteMapper;
 		$this->shareMapper = $shareMapper;
 		$this->poll = $poll;
 	}
@@ -112,9 +118,31 @@ class Acl implements JsonSerializable {
 	 * @NoAdminRequired
 	 * @return string
 	 */
+	public function getDisplayName() {
+		$this->userManager = \OC::$server->getUserManager();
+
+		if (\OC::$server->getUserManager()->get($this->userId) instanceof IUser) {
+			return \OC::$server->getUserManager()->get($this->userId)->getDisplayName();
+		} else {
+			return $this->userId;
+		}
+	}
+
+	/**
+	 * @NoAdminRequired
+	 * @return string
+	 */
 	public function setUserId($userId): Acl {
 		$this->userId = $userId;
 		return $this;
+	}
+
+	/**
+	* @NoAdminRequired
+	* @return string
+	*/
+	public function getLoggedIn() {
+		return \OC::$server->getUserSession()->isLoggedIn();
 	}
 
 	/**
@@ -168,11 +196,15 @@ class Acl implements JsonSerializable {
 	public function getAllowView(): bool {
 		return (
 			   $this->getIsOwner()
-			|| $this->getIsAdmin()
-			|| ($this->getGroupShare() && !$this->poll->getDeleted())
-			|| ($this->getPersonalShare() && !$this->poll->getDeleted())
-			|| $this->poll->getAccess() !== 'hidden'
-			);
+			|| ($this->getIsAdmin() && $this->poll->getAdminAccess())
+			|| !$this->poll->getDeleted() && (
+				   $this->getUserHasVoted()
+				|| $this->getGroupShare()
+				|| $this->getPersonalShare()
+				|| $this->getPublicShare()
+				|| ($this->poll->getAccess() !== 'hidden' && !$this->getPublicShare())
+			)
+		);
 	}
 
 	/**
@@ -193,11 +225,36 @@ class Acl implements JsonSerializable {
 	 * @NoAdminRequired
 	 * @return bool
 	 */
+	public function getUserHasVoted(): bool {
+		return count(
+			$this->voteMapper->findParticipantsVotes($this->getPollId(), $this->getUserId())
+		);
+	}
+
+	/**
+	 * @NoAdminRequired
+	 * @return bool
+	 */
 	public function getPersonalShare(): bool {
 
 		return count(
 			array_filter($this->shareMapper->findByPoll($this->getPollId()), function($item) {
 				if (($item->getType() === 'user' || $item->getType() === 'external') && $item->getUserId() === $this->getUserId()) {
+					return true;
+				}
+			})
+		);
+	}
+
+	/**
+	 * @NoAdminRequired
+	 * @return bool
+	 */
+	public function getPublicShare(): bool {
+
+		return count(
+			array_filter($this->shareMapper->findByPoll($this->getPollId()), function($item) {
+				if ($item->getType() === 'public' && $item->getToken() === $this->getToken()) {
 					return true;
 				}
 			})
@@ -221,9 +278,11 @@ class Acl implements JsonSerializable {
 	 */
 	public function getAllowVote(): bool {
 		if (
-			   $this->getAllowView()
+			   ($this->getAllowView() || $this->getFoundByToken())
 			&& !$this->getExpired()
 			&& !$this->poll->getDeleted()
+			&& $this->userId
+
 		) {
 			return true;
 		} else {
@@ -345,6 +404,8 @@ class Acl implements JsonSerializable {
 	public function jsonSerialize(): array {
 		return	[
 			'userId'            => $this->getUserId(),
+			'displayName'       => $this->getDisplayName(),
+			'loggedIn'			=> $this->getLoggedIn(),
 			'pollId'            => $this->getPollId(),
 			'token'             => $this->getToken(),
 			'isOwner'           => $this->getIsOwner(),
@@ -355,8 +416,10 @@ class Acl implements JsonSerializable {
 			'allowEdit'         => $this->getAllowEdit(),
 			'allowSeeUsernames' => $this->getAllowSeeUsernames(),
 			'allowSeeAllVotes'  => $this->getAllowSeeAllVotes(),
+			'userHasVoted'		=> $this->getUserHasVoted(),
 			'groupShare'        => $this->getGroupShare(),
 			'personalShare'     => $this->getPersonalShare(),
+			'publicShare'     	=> $this->getPublicShare(),
 			'foundByToken'      => $this->getFoundByToken(),
 			'accessLevel'       => $this->getAccessLevel()
 		];