+++ src/js/components/PollList/PollListItem.vue
@@ -67,12 +67,12 @@
 				{{ t('polls', 'Clone poll') }}
 			</ActionButton>
 
-			<ActionButton v-if="poll.owner === OC.getCurrentUser().uid && !poll.deleted" icon="icon-delete" @click="switchDeleted()">
-				{{ t('polls', 'Delete poll') }}
+			<ActionButton v-if="poll.allowEdit && !poll.deleted" icon="icon-delete" @click="switchDeleted()">
+				{{ (poll.isAdmin) ? t('polls', 'Delete poll as admin') : t('polls', 'Delete poll') }}
 			</ActionButton>
 
-			<ActionButton v-if="poll.owner === OC.getCurrentUser().uid && poll.deleted" icon="icon-history" @click="switchDeleted()">
-				{{ t('polls', 'Restore poll') }}
+			<ActionButton v-if="poll.allowEdit && poll.deleted" icon="icon-history" @click="switchDeleted()">
+				{{ (poll.isAdmin) ? t('polls', 'Restore poll as admin') : t('polls', 'Restore poll') }}
 			</ActionButton>
 		</Actions>
 
@@ -260,7 +260,7 @@ export default {
 
 	.owner {
 		flex: 0 0 auto;
-		width: 130px;
+		width: 230px;
 		overflow: hidden;
 		white-space: nowrap;
 		text-overflow: ellipsis;