+++ src/js/views/NotFound.vue
@@ -21,59 +21,24 @@
   -->
 
 <template>
-	<AppSidebar ref="sideBar" :title="t('polls', 'Details')" @close="$emit('closeSideBar')">
-		<UserDiv slot="primary-actions" :user-id="poll.owner" :description="t('polls', 'Owner')" />
-
-		<AppSidebarTab :name="t('polls', 'Comments')" icon="icon-comment">
-			<Comments />
-		</AppSidebarTab>
-	</AppSidebar>
+	<AppContent>
+		<div id="emptycontent">
+			<div id="emptycontent-icon" class="icon-search" />
+			<h2>{{ t('polls', 'The poll does not exist') }}</h2>
+			<p class="emptycontent-additional">
+				{{ t('polls', 'Enter a poll or start a new one.') }}
+			</p>
+		</div>
+	</AppContent>
 </template>
 
 <script>
-import { AppSidebar, AppSidebarTab } from '@nextcloud/vue'
-
-import Comments from '../Comments/Comments'
-import { mapState } from 'vuex'
+import { AppContent } from '@nextcloud/vue'
 
 export default {
-	name: 'SideBarOnlyComments',
+	name: 'NotFound',
 	components: {
-		Comments,
-		AppSidebar,
-		AppSidebarTab
-	},
-
-	computed: {
-		...mapState({
-			poll: state => state.poll,
-			acl: state => state.acl
-		})
+		AppContent
 	}
-
 }
-
 </script>
-
-<style scoped lang="scss">
-
-	ul {
-		& > li {
-			margin-bottom: 30px;
-			& > .comment-item {
-				display: flex;
-				align-items: center;
-
-				& > .date {
-					right: 0;
-					top: 5px;
-					opacity: 0.5;
-				}
-			}
-			& > .message {
-				margin-left: 44px;
-				flex: 1 1;
-			}
-		}
-	}
-</style>